significantly updated code and formatting
moved operations into separate functions for readability and testability: character prediction, character decoding, etc.

made lots of tweaks to how the bullets are generated. Learned that it is very essential to pass in the previous prediction tensor into model input, as
the model carries state information and gets lost if a different tensor (identified with id parameter) is passed into it.

If you want to add suggestions mid-run, it therefore makes more sense to restart the model and push into it all previous outputs + the new suggestion.

I ended up making a 'free run' and a 'same prefix' version of the generator, I think they seem to output bullets coherently. the free run method seems to be cleaner.

Also added some bulma css to prettify the site.